Add mode to search and view specific reflection
A frequent usecase of psalm is to match a reflection identified by
real-world measurement to a reflection predicted by a simulation. This
commit adds a mode `--search-itd` that finds the closest reflection to
the requested ITD. The user can optionally also specify `--gain`, in
which case a reflection with close to the target gain will be preferred.

This commit also displays additional data about the gain contribution of
the angle of emergence of a ray that results in a reflection.